simple_app/requirement.txt
Flask
werkzeug
flask-session
flask_cors
pytz
requests
flask-pymongo
dnspython